+++ src/common/event/FullyLinkedEvent.ts
@@ -1,8 +1,8 @@
 import { Edge } from "../item/Edge/types/Edge";
-import { Node } from "../item/Node/types/Node";
+import { Node, ProcessedNode } from "../item/Node/types/Node";
 
-export interface FullyLinkedEvent<NodeType, EdgeType, SpecificFullyLinkedEventInfo> {
-  item: Node<NodeType> | Edge<EdgeType> | null;
+export interface FullyLinkedEvent<NodeType, EdgeType, SpecificFullyLinkedEventInfo, GlobalNodePropsType> {
+  item: ProcessedNode<NodeType, GlobalNodePropsType> | Edge<EdgeType> | null;
   itemType: "node" | "edge" | null;
   info: SpecificFullyLinkedEventInfo;
 }