+++ src/common/item/node/functions/setUpCreateEdgeOnAnchorDragging.ts
@@ -2,7 +2,7 @@ import * as d3 from "d3";
 import { curveBumpX } from "d3";
 import { Disposer } from "../../../disposer/Disposer";
 import { CreateNewEdgeStateMaintainer } from "../../edge/stateMaintainers/CreateNewEdgeStateMaintainer";
-import { InternalNode } from "./../types/Node";
+import { ProcessedNode } from "./../types/Node";
 import { addDisposableEventListener } from "../../../event/addEventListener";
 import { CanvasZoomAndTransformMaintainer } from "../../canvas/stateMaintainers/CanvasZoomAndTransformMaintainer";
 import { Edge } from "../../edge/types/Edge";
@@ -11,24 +11,24 @@ import { FullyLinkedEvent } from "../../../event/FullyLinkedEvent";
 import { dispatchFullyLinkedEvent } from "../../../event/dispatchFullyLinkedEvent";
 import { FullyLinkedEventEnum } from "../../../event/FullyLinkedEventEnum";
 
-interface CreateEdgeOnAnchorDraggingParams<NodeType, EdgeType> {
+interface CreateEdgeOnAnchorDraggingParams<NodeType, EdgeType, GlobalNodePropsType> {
   anchorEndElem: HTMLDivElement;
   anchorStartElem: HTMLDivElement;
-  node: InternalNode<NodeType>;
+  node: ProcessedNode<NodeType, GlobalNodePropsType>;
   internalSVGElement: SVGSVGElement | undefined;
   edgePlaceholderId: string;
   createNewEdgeStateMaintainer: CreateNewEdgeStateMaintainer;
   canvasZoomLevelMaintainer: CanvasZoomAndTransformMaintainer;
   disposer: Disposer;
   container: HTMLElement;
-  nodeMapById: Map<string, InternalNode<NodeType>>;
+  nodeMapById: Map<string, ProcessedNode<NodeType, GlobalNodePropsType>>;
   edgeMapById: Map<string, Edge<EdgeType>>;
   edgeListMapByNodeId: Map<string, Edge<EdgeType>[]>;
 }
 
 /** When user drags an "end anchor" element, a new placeholder edge should be created.
  * The placeholder edge should become a real edge when the user drops the end of the edge onto another node's "start anchor" */
-export function setUpCreateEdgeOnAnchorDragging<NodeType, EdgeType>({
+export function setUpCreateEdgeOnAnchorDragging<NodeType, EdgeType, GlobalNodePropsType>({
   anchorEndElem,
   anchorStartElem,
   node,
@@ -41,7 +41,7 @@ export function setUpCreateEdgeOnAnchorDragging<NodeType, EdgeType>({
   edgeMapById: edgesMapById,
   edgeListMapByNodeId: edgesMapByNodeId,
   nodeMapById: nodesMapById,
-}: CreateEdgeOnAnchorDraggingParams<NodeType, EdgeType>) {
+}: CreateEdgeOnAnchorDraggingParams<NodeType, EdgeType, GlobalNodePropsType>) {
   // Positions
   let dragStartX: number;
   let dragStartY: number;
@@ -50,7 +50,7 @@ export function setUpCreateEdgeOnAnchorDragging<NodeType, EdgeType>({
 
   const anchorEndMouseDownListener = (e: PointerEvent) => {
     // First, dispatch edgeCreationStart event
-    const beforeEdgeCreationParams: FullyLinkedEvent<NodeType, EdgeType, {}> = {
+    const beforeEdgeCreationParams: FullyLinkedEvent<NodeType, EdgeType, {}, GlobalNodePropsType> = {
       item: null,
       itemType: "edge",
       info: {},